seperated headless node into own file and run scripts
